Fix convert of md-uv380 codeplug to md-2017
In this case, we were trying to delete button definitions.  The
number of button definitions is fixed.  When converting to a
codeplug that has non-variable-count records (like button definitions)
and the number of records is less than supported by the codeplug type,
we create the missing records.

Thanks to Eduardo Nunes, CT4AJ, for the bug report.