Filter records referenced by other records' fields
We maintain a static list of all possible record types that have
fields that reference a given record type.  Traversing this list
fields if the codeplug doesn't have a specific record type.  We
now filter out those record types that do not exist in the codeplug.

Thanks to Mike, KB2FX, for reporting the bug.